Move apiAlert
