+++ src/main/java/org/opentripplanner/api/model/ApiWalkStep.java
@@ -1,7 +1,5 @@
 package org.opentripplanner.api.model;
 
-import org.opentripplanner.api.model.alertpatch.ApiAlert;
-
 import java.util.List;
 
 /**
